pyproject.toml
[tool.poetry]
name = "gpt-app"
version = "0.1.0"
description = ""
authors = ["emmc15 <https://github.com/emmc15>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.228"
gpt4all = "^1.0.3"
chromadb = "^0.3.26"
llama-cpp-python = "^0.1.70"
urllib3 = "^2.0.3"
pymupdf = "^1.22.5"
python-dotenv = "^1.0.0"
unstructured = "^0.8.0"
extract-msg = "^0.41.5"
tabulate = "^0.9.0"
pandoc = "^2.3"
pypandoc = "^1.11"
tqdm = "^4.65.0"
gunicorn = "^20.1.0"
streamlit = "^1.24.1"
python-multipart = "^0.0.6"
torch = "!=2.0.1"
sentence-transformers = "^2.2.2"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
